refactor: #42 game page refactor
