+++ src/app/game/data/games.ts
@@ -2,7 +2,6 @@ import { PongGameWindowComponent } from '../games/pong/pong.component';
 import { Game } from '../models/game.class';
 import { TictactoeGameWindowComponent } from '../games/tictactoe/tictactoe.component';
 import { Player } from '../models/player.class';
-import { PlayerSourceType } from '../models/player-source-type.enum';
 
 export const games: Record<string, Game> = {
   pong: new Game('pong', PongGameWindowComponent, [
